Preliminary removal of integrated login code from wp-integration-class to internalise it into wordpress -- in login-integrator.php. Lots of work still to do, expect breakages in installer/usermapper/etc.
git-svn-id: http://www.wp-united.com/svn/trunk@437 c1c82642-cce3-494f-bf0e-9d5eb4df7846

+++ root/includes/acp/acp_wp_united.php
@@ -1475,6 +1475,11 @@ class acp_wp_united {
 						$wpUtdInt->enter_wp_integration();
 						eval($wpUtdInt->exec());  
 						
+						if(!function_exists('update_option')) {
+							die('ERROR: WordPress could not be run. Please check it is installed properly and that you provided the correct paths.');
+						}
+						
+						
 						// reset the Blog Uri
 						if (!empty($wpSettings['wpUri'])) {
 							update_option('home', $wpSettings['wpUri']);
